Implemented profiles

+++ src/main/java/com/narola/testspringboot/medicine/repository/MedicineRepository.java
@@ -3,6 +3,7 @@ package com.narola.testspringboot.medicine.repository;
 import com.narola.testspringboot.model.Medicine;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MedicineRepository extends CrudRepository<Medicine,Long> {
+public interface MedicineRepository extends CrudRepository<Medicine,Integer> {
+
 
 }
